Refactor dump and restore to reuse
