+++ ydb/public/lib/ydb_cli/dump/util/external_table_utils.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <util/generic/fwd.h>
+
+namespace Ydb::Table {
+    class DescribeExternalTableResult;
+}
+
+namespace NYql {
+    class TIssues;
+}
+
+namespace NYdb::NDump {
+    TString BuildCreateExternalTableQuery(const Ydb::Table::DescribeExternalTableResult& description);
+
+    bool RewriteCreateExternalTableQuery(
+        TString& query,
+        const TString& dbPath,
+        NYql::TIssues& issues);
+}
