+++ ydb/public/lib/ydb_cli/dump/util/ya.make
@@ -4,6 +4,9 @@ SRCS(
     query_utils.cpp
     util.cpp
     view_utils.cpp
+    replication_utils.cpp
+    external_data_source_utils.cpp
+    external_table_utils.cpp
 )
 
 PEERDIR(
